Pull out helper methods
