+++ lib/foods.js
@@ -1,24 +1,10 @@
-const url = 'https://quantified-self-api-ad.herokuapp.com/'
+const helperFile = require('./helper.js')
 
 const getFoods = () =>  {
-  fetch(`${url}api/v1/foods`)
-  .then((response) => handleResponse(response))
+  fetch(`${helperFile.url}api/v1/foods`)
+  .then((response) => helperFile.handleResponse(response))
   .then((foods) => getEachFood(foods))
   .catch((error) => console.error ({ error }))
-}
-
-const handleResponse = (response) =>  {
-  return response.json()
-  .then(json => {
-    if (!response.ok) {
-      const error = {
-        status: response.status,
-        statusText: response.statusText,
-      }
-      return Promise.reject(error);
-    }
-    return json
-  })
 }
 
 const getEachFood = (foodData) => {
@@ -69,7 +55,7 @@ const addNewFood = (event) => {
       newCalories = $("#new-food-calorie-count").val(),
       body = {food: {name: newName, calories: newCalories}}
   makeRequest('POST', '', body)
-  .then((response) => handleResponse(response))
+  .then((response) => helperFile.handleResponse(response))
   .then((food) => renderFood(food))
   .catch((error) => console.error({error}))
   $("#new-food-form").trigger('reset')
@@ -96,7 +82,7 @@ $("#food-table").keypress((event) => {
 })
 
 const makeRequest = (method, id, body) => {
-  return fetch(`${url}api/v1/foods/${id}`, {
+  return fetch(`${helperFile.url}api/v1/foods/${id}`, {
     method: `${method}`,
     headers: {
       "Content-Type": "application/json"
@@ -112,7 +98,7 @@ const editFood = (event) => {
   let calorieContent = row.childNodes[3].textContent
   let body = {food: { name: nameContent, calories: calorieContent}}
   makeRequest('PATCH', id, body)
-  .then((response) => handleResponse(response))
+  .then((response) => helperFile.handleResponse(response))
   .catch((error) => console.error({error}))
 }
 
@@ -136,4 +122,4 @@ const filterFoods = () => {
 $('#search-term').on('keyup', filterFoods)
 
 export * from './foods.js'
-module.exports = {getFoods, filterFoods, renderFood, url, handleResponse, getPage}
+module.exports = {getFoods, renderFood, getPage}
