+++ lib/index.js
@@ -2,28 +2,13 @@ import * as Style from '../main.css'
 import * as Foods from './foods.js'
 
 const foodFile = require('./foods.js')
-
-const url = 'https://quantified-self-api-ad.herokuapp.com/'
+const helperFile = require('./helper.js')
 
 const getMeals = () => {
-  fetch(`${url}api/v1/meals`)
-  .then((response) => handleResponse(response))
+  fetch(`${helperFile.url}api/v1/meals`)
+  .then((response) => helperFile.handleResponse(response))
   .then((meals) => getEachMeal(meals))
   .catch((error) => console.error({ error }))
-}
-
-const handleResponse = (response) => {
-  return response.json()
-  .then(json => {
-    if (!response.ok) {
-      const error = {
-        status: response.status,
-        statusText: response.statusText,
-      }
-      return Promise.reject(error);
-    }
-    return json
-  })
 }
 
 const getEachMeal = (mealData) => {
@@ -121,7 +106,7 @@ $("#meal-buttons").on("click", function(event) {
     checkedFoodIds.push(checkedFoods[i].closest('tr').dataset.food)
   }
   checkedFoodIds.forEach(foodId => {
-    fetch(`${foodFile.url}api/v1/meals/${mealId}/foods/${foodId}`, {
+    fetch(`${helperFile.url}api/v1/meals/${mealId}/foods/${foodId}`, {
       method: 'POST',
       headers: {
         "Content-Type": "application/json"
@@ -129,8 +114,8 @@ $("#meal-buttons").on("click", function(event) {
     })
     .catch((error) => console.error({error}))
   })
-  fetch(`${foodFile.url}api/v1/meals/${mealId}/foods`)
-  .then(response => foodFile.handleResponse(response))
+  fetch(`${helperFile.url}api/v1/meals/${mealId}/foods`)
+  .then(response => helperFile.handleResponse(response))
   .then(meal => renderMeals(meal))
 })
